backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sentence-transformers
faiss-cpu
spacy
pdfplumber
python-docx
pillow
pytesseract
paddleocr
nltk
requests
openai
uvicorn
pydantic_settings
python-multipart
paddleocr
groq
python-dotenv